+++ src/components/Form/Checkbox/Checkbox.tsx
@@ -1,5 +1,5 @@
 // FRAMEWORK ===========================================================================================================
-import React from "react";
+import React, { useMemo } from "react";
 
 // FICTOAN =============================================================================================================
 import { Element } from "../../Element/Element";
@@ -15,10 +15,23 @@ export type CheckboxElementType = HTMLInputElement;
 export type CheckboxProps = Omit<BaseInputComponentProps<CheckboxElementType>, "as">;
 
 // COMPONENT ///////////////////////////////////////////////////////////////////////////////////////////////////////////
-export const Checkbox = React.forwardRef(({ ...props }: CheckboxProps, ref: React.Ref<CheckboxElementType>) => {
-    return (
-        <Element<CheckboxElementType> as="div" data-checkbox ref={ref}>
-            <BaseInputComponent as="input" type="checkbox" {...props} />
-        </Element>
-    );
-});
+export const Checkbox = React.forwardRef(
+    ({ id, name, value, ...props }: CheckboxProps, ref: React.Ref<CheckboxElementType>) => {
+        // Use ID as default for name and value if they’re not provided
+        const derivedName = useMemo(() => name || id, [name, id]);
+        const derivedValue = useMemo(() => value || id, [value, id]);
+
+        return (
+            <Element<CheckboxElementType> as="div" data-checkbox ref={ref}>
+                <BaseInputComponent
+                    as="input"
+                    type="checkbox"
+                    id={id}
+                    name={derivedName}
+                    value={derivedValue}
+                    {...props}
+                />
+            </Element>
+        );
+    }
+);