+++ src/components/Form/RadioButton/RadioTabGroup.tsx
@@ -1,25 +1,42 @@
-import React from "react";
+// FRAMEWORK ===========================================================================================================
+import React, { useMemo } from "react";
 
+// FICTOAN =============================================================================================================
 import { Div } from "../../Element/Tags";
 import { BaseInputComponent } from "../BaseInputComponent/BaseInputComponent";
-import { RadioTabGroupProps, RadioGroupProps, RadioButtonElementType } from "./constants";
 
+// STYLES ==============================================================================================================
 import "./radio-tab-group.css";
 
-const RadioTabGroupOptions = ({ options, defaultValue, value, required, ...props }: RadioGroupProps) => {
+// TYPES ===============================================================================================================
+import { RadioTabGroupProps, RadioGroupProps, RadioButtonElementType } from "./constants";
+
+// COMPONENT ///////////////////////////////////////////////////////////////////////////////////////////////////////////
+const RadioTabGroupOptions = ({ id, name, options, defaultValue, value, required, ...props }: RadioGroupProps) => {
+    // Use ID as default for name if not provided
+    const derivedName = useMemo(() => name || id, [name, id]);
+
     return (
-        <Div data-radio-tab-group name={props.name} required={required}>
-            {options.map((option) => (
-                <React.Fragment key={option.id}>
-                    <input
-                        type="radio"
-                        {...props}
-                        {...option}
-                        checked={value === option.value}
-                    />
-                    <label htmlFor={option.id}>{option.label}</label>
-                </React.Fragment>
-            ))}
+        <Div data-radio-tab-group name={derivedName} required={required}>
+            {options.map((option, index) => {
+                const { id: optionId, ...optionProps } = option;
+                // Derive option id if not provided
+                const finalId = optionId || `${id}-option-${index}`;
+
+                return (
+                    <React.Fragment key={finalId}>
+                        <input
+                            type="radio"
+                            {...props}
+                            {...optionProps}
+                            id={finalId}
+                            name={derivedName}
+                            checked={value === option.value}
+                        />
+                        <label htmlFor={finalId}>{option.label}</label>
+                    </React.Fragment>
+                );
+            })}
         </Div>
     );
 };